+++ fabric-samples/chaincode/fabcar/java/src/main/java/iReceptorPlus/Blockchain/iReceptorChain/VotingRoundStateMachine/Exceptions/EntityDoesNotHaveEnoughReputationToPerformAction.java
@@ -2,9 +2,9 @@ package iReceptorPlus.Blockchain.iReceptorChain.VotingRoundStateMachine.Exceptio
 
 public class EntityDoesNotHaveEnoughReputationToPerformAction extends TraceabilityInfoStateMachineException
 {
-    Long reputationOfEntity;
+    Double reputationOfEntity;
 
-    Long necessaryReputation;
+    Double necessaryReputation;
 
     /**
      * Constructor for this class that receives the message as parameter.
@@ -12,19 +12,19 @@ public class EntityDoesNotHaveEnoughReputationToPerformAction extends Traceabili
      *
      * @param message The string message that will be set as the message for the exception that occurred.
      */
-    public EntityDoesNotHaveEnoughReputationToPerformAction(String message, Long reputationOfEntity, Long necessaryReputation)
+    public EntityDoesNotHaveEnoughReputationToPerformAction(String message, Double reputationOfEntity, Double necessaryReputation)
     {
         super(message);
         this.reputationOfEntity = reputationOfEntity;
         this.necessaryReputation = necessaryReputation;
     }
 
-    public Long getReputationOfEntity()
+    public Double getReputationOfEntity()
     {
         return reputationOfEntity;
     }
 
-    public Long getNecessaryReputation()
+    public Double getNecessaryReputation()
     {
         return necessaryReputation;
     }
